SecSplitLLM/split-fed-learning/requirements.txt
codecarbon
flwr
datasets
matplotlib
pandas
protobuf
scikit-learn
seaborn
tensorboard
tensorflow
torch
torchaudio 
torchmetrics
torchvision 
tqdm
transformers
wandb
grpcio-tools
opacus
GPUtils